chore: Refactor AddVendedorModal.js and HomePage.js
This commit refactors the code in AddVendedorModal.js and HomePage.js. It removes the unused import of getVendedores from idbService in AddVendedorModal.js and replaces it with a placeholder logic comment. Additionally, it removes the ConfirmDeleteModal component and replaces it with a Dialog component in HomePage.js. These changes improve code organization and remove unnecessary dependencies.