feat: Refactor code in AuthContext.js and add localStorage support
This commit refactors the code in AuthContext.js to include the use of localStorage for storing user data. It also adds a useEffect hook to redirect the user to the login page if there is no user data present. These changes improve the user authentication functionality and provide a better user experience.